Major Updates: Error Handling
Added a basic addition command, a class for making the terminal a little more colorful, as well as ERROR HANDLING, and a LOT of it. It's something I've wanted to do for some time, so I decided to just go ahead and add it right away.

TODO: Add a custom error message for certain commands, I need to figure out how to do this.